Falta nao-idempotencia

+++ rpc/rpcgen/bank.h
@@ -28,7 +28,7 @@ typedef struct deposita_1_argument deposita_1_argument;
 
 struct saca_1_argument {
 	int arg1;
-	float arg2;
+	int arg2;
 };
 typedef struct saca_1_argument saca_1_argument;
 
@@ -55,8 +55,8 @@ extern  float * saldo_1_svc(int , struct svc_req *);
 extern  int * deposita_1(int , float , CLIENT *);
 extern  int * deposita_1_svc(int , float , struct svc_req *);
 #define SACA 7
-extern  int * saca_1(int , float , CLIENT *);
-extern  int * saca_1_svc(int , float , struct svc_req *);
+extern  int * saca_1(int , int , CLIENT *);
+extern  int * saca_1_svc(int , int , struct svc_req *);
 #define ASSINATURA 8
 extern  int * assinatura_1(CLIENT *);
 extern  int * assinatura_1_svc(struct svc_req *);
